wip: model features
