+++ client/src/features/model/index.ts
@@ -1,11 +1,23 @@
-// UI Components - Tabs
-export { ModelOverviewTab } from "./ui/tabs/ModelOverviewTab";
-export { ModelInferenceTab } from "./ui/tabs/ModelInferenceTab";
-export { LayerFlowVisualization } from "./ui/tabs/LayerFlowVisualization";
-export { NeuralNetworkVisualization } from "./ui/tabs/NeuralNetworkVisualization";
+// Types
+export type { Model, ModelFilters, ModelListState, ModelCardProps, ModelDetailProps } from './types';
+
+// Hooks
+export { useModelFilters } from './hooks/useModelFilters';
 
 // UI Components
-export { ModelCard } from "./ui/components/ModelCard";
+export { 
+  ModelSearchFilters, 
+  ModelList, 
+  ModelDetailHeader 
+} from './ui';
+
+// Legacy exports (기존 컴포넌트들 유지)
+export { ModelCard } from './ui/components/ModelCard';
+export { ModelOverviewTab, ModelInferenceTab } from './ui/tabs';
+
+// UI Components - Tabs
+export { LayerFlowVisualization } from "./ui/tabs/LayerFlowVisualization";
+export { NeuralNetworkVisualization } from "./ui/tabs/NeuralNetworkVisualization";
 
 // Re-export types for convenience
 export type { ModelObject, Layer } from "@/shared/api/graphql/modelGraphQLService";
