+++ client/src/features/model/ui/index.ts
@@ -1,0 +1,3 @@
+export { ModelSearchFilters } from './ModelSearchFilters/ModelSearchFilters';
+export { ModelList } from './ModelList/ModelList';
+export { ModelDetailHeader } from './ModelDetailHeader/ModelDetailHeader';
